backend/flask-api/requirements.txt
Flask
tensorflow==2.7
numpy
pillow
requests
Keras-Applications